+++ PyFiberModes/solver/__init__.py
@@ -1,8 +1,8 @@
 """Numerical solvers for fiber cutoff and effective-index equations."""
 
-from . import mlsif as mlsif
-from . import ssif as ssif
-from . import tlsif as tlsif
+from . import multilayer as multilayer
+from . import three_layer as three_layer
+from . import two_layer as two_layer
 from .protocols import CutoffSolver as CutoffSolverProtocol
 from .protocols import EffectiveIndexSolver, RadialFieldSolver
 from .results import SolverResult
@@ -12,7 +12,7 @@ __all__ = [
     "EffectiveIndexSolver",
     "RadialFieldSolver",
     "SolverResult",
-    "mlsif",
-    "ssif",
-    "tlsif",
+    "multilayer",
+    "three_layer",
+    "two_layer",
 ]