+++ PyFiberModes/solver/mlsif/__init__.py
@@ -1,3 +1,0 @@
-"""Multilayer step-index fiber solvers."""
-
-from .neff import NeffSolver as NeffSolver